+++ src/hmmv2/sample.rs
@@ -2,9 +2,9 @@
 //! Sampling emissions from the PHMMModel
 //!
 use super::common::{PHMMEdge, PHMMModel, PHMMNode};
-use crate::hmm::params::PHMMParams;
 use crate::prob::Prob;
 pub use petgraph::graph::{EdgeIndex, NodeIndex};
+use picker::{pick_ins_emission, pick_match_emission, pick_with_prob};
 use rand::prelude::*;
 use rand_xoshiro::Xoshiro256PlusPlus;
 pub mod picker;
@@ -27,11 +27,6 @@ pub enum State {
 ///
 #[derive(Debug, Copy, Clone)]
 pub struct Emission(pub Option<u8>);
-
-///
-/// Array of valid DNA bases
-///
-const BASES: [u8; 4] = [b'A', b'C', b'G', b'T'];
 
 impl<N: PHMMNode, E: PHMMEdge> PHMMModel<N, E> {
     ///
@@ -58,7 +53,7 @@ impl<N: PHMMNode, E: PHMMEdge> PHMMModel<N, E> {
 
         history
     }
-    fn sample_init<R: Rng>(&self, rng: &mut R) -> (State, Emission) {
+    fn sample_init<R: Rng>(&self, _rng: &mut R) -> (State, Emission) {
         (State::MatchBegin, Emission(None))
     }
     fn sample_step<R: Rng>(&self, rng: &mut R, now: State) -> (State, Emission) {
@@ -200,45 +195,6 @@ impl<N: PHMMNode, E: PHMMEdge> PHMMModel<N, E> {
             Some(pick_with_prob(rng, &choices))
         }
     }
-}
-
-///
-/// pick randomly from the choices with its own probability.
-///
-pub fn pick_with_prob<R: Rng, T: Copy>(rng: &mut R, choices: &[(T, Prob)]) -> T {
-    choices
-        .choose_weighted(rng, |item| item.1.to_value())
-        .unwrap()
-        .0
-}
-
-///
-/// Pick an emission at `Ins` state, according to the PHMM params.
-/// It selects uniformaly random on `ACGT`
-///
-pub fn pick_ins_emission<R: Rng>(rng: &mut R, param: &PHMMParams) -> Emission {
-    let base = *BASES
-        .choose_weighted(rng, |_base| param.p_random.to_value())
-        .unwrap();
-    Emission(Some(base))
-}
-
-///
-/// Pick an emission at `Match` state (with a state emission), according to the PHMM params.
-/// It selects a state emission with probability `p_match`, and other emissions with probability
-/// `p_mismatch`.
-///
-pub fn pick_match_emission<R: Rng>(rng: &mut R, emission: u8, param: &PHMMParams) -> Emission {
-    let base = *BASES
-        .choose_weighted(rng, |&base| {
-            if base == emission {
-                param.p_match.to_value()
-            } else {
-                param.p_mismatch.to_value()
-            }
-        })
-        .unwrap();
-    Emission(Some(base))
 }
 
 #[cfg(test)]
